contributions/containerization-rychu-prime/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
asyncpg==0.30.0
redis==6.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
prometheus-client==0.22.1
python-multipart==0.0.20
httpx==0.28.1
